deleting case document

+++ src/application/usecases/CaseDocuments/implementations/DeleteCaseDocument.ts
@@ -7,10 +7,14 @@ export class DeleteCaseDocumentsUseCase implements IDeleteCaseDocument {
     private _caseDocumentsRepo: ICaseDocumentsRepo,
     private _cloudinaryService: ICloudinaryService
   ) {}
-  async execute(input: string): Promise<void> {
-    const existingCaseDocuments = await this._caseDocumentsRepo.findById(input);
+  async execute(input: { documentId: string; userId: string }): Promise<void> {
+    const existingCaseDocuments = await this._caseDocumentsRepo.findById(
+      input.documentId
+    );
     if (!existingCaseDocuments) throw new Error("case documents not found");
-    await this._caseDocumentsRepo.delete(input);
+    if (existingCaseDocuments.uploadedBy !== input.userId)
+      throw new Error("Unable to delete others uploaded file");
+    await this._caseDocumentsRepo.delete(input.documentId);
     await this._cloudinaryService.deleteFile(
       existingCaseDocuments.document.url
     );
